fuse: fix getattr/ opendir for dirs missing in backing FS
Don't assume that directories will always be available in the backing
FS; try the fstree.  This doesn't fix readdir, so missing directories
won't be shown in the 'ls' output, but can be opened.  Unfortunately,
drs are inserted directly into the fsroot -- this needs to be changed
to some recursive format to ease lookups.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>